No prestige from partition cb's
Yeah, doing this would actually destroy the prestige system as it gave a ridiculous
amount of prestige for adding them. I think it was giving 50% of the nation's total
prestige, instead of 50% of the normal prestige that the cb would yield. This was a
bit of a problem, I discovered in my germany game, where I went from about 1500ish
to 12000 after annexing three countries.