Cores will now happen more quickly
srsly tho, why wouldn't you get more benefit if the province is primary culture and not just an accepted culture. like wtf guys.

also 15 years is a really long time, so I reduced it to 12. 15 years might make sense irl, but with the mean time to happen thing, that just means it'll probably take up to 60 years for a province with 100% primary culture to become a core, which is just insane, especially if there are not even other cores there.

So now, there is:
a 200% bonus for being majority primary culture & noiit just accepted