Updated country colors; Made rebels fully black
Rebel-controlled territories are now shaded with full black, instead of
the previous mid-grey. In my experience, the previous rebel color was
difficult to see against some countries' colors, while this seems to
contrast fairly well against all colors, while still being able to be
seen. One downside is that for small, darker colors (like Montenegro), the
color might not be very visible. But this might just be due to the fact
that the country is so tiny.

USA is light pink cuz it was funny

Prussia has been reverted to its yellow color of DLCs past, cuz it was
funny.

France is a disgusting purple also cuz it was funny too.