Relaxed conquest cb requirements
Old requirement:
- Must have 1 state

New requirement:
- One of the following:
  - 1 state
  - fewer than 5 provinces (excl.) (i.e. <=4 provinces)
  - fewer than 250,000 total population

This means that small countries (either through land or population) can
be more easily conquested. This is especially useful when a country has
like 3 provinces split across 2 or 3 states, which means you'll have to
go to war 2 or 3 times just to fully annex them, which is kinda stupid.