Add the Lemko republic (#29)
* Created Lemko Republic

It's basically just copied Krakow, but with a few different names and a
hideous color.

* Added (temporary) flags

Only the LMK and LMK_republic files are permanent, as these were the
actual flags of the country. The others I will have to customize myself
later. These are just placeholders to get the game to load.

* Added core for Lemko to prov 704

This is where they were, like even the shape is really close to how the
country actually looked. How convenient! It's like Lemko was supposed to
be in the game!

* Added localisations for the name of LMK

* Added history for Lemko

Lemko will have Ukranian as its primary culture, as per what wikipedia
says Lemkos are. And polish is an accepted culture because they're in
poland.

* Removed Polish accepted culture from LMK

Upon further research, the peoples are not very similar at all and thus
the Lemkos shouldn't accept the Poles.

* Fixed the flags

Apparently Paint.NET doesn't save by file extension

* Made Lemko's color more intereseting

* LMK now has an upperhouse

tbh, I thought it would just be figured out automatically when released,
but apparently not.

* LMK parties now have names

They're just the Ukrainian party names, but with "Lemko" instead of
"Ukrainian"

* Created Lemko Republic

It's basically just copied Krakow, but with a few different names and a
hideous color.

* Added (temporary) flags

Only the LMK and LMK_republic files are permanent, as these were the
actual flags of the country. The others I will have to customize myself
later. These are just placeholders to get the game to load.

* Added core for Lemko to prov 704

This is where they were, like even the shape is really close to how the
country actually looked. How convenient! It's like Lemko was supposed to
be in the game!

* Added localisations for the name of LMK

* Added history for Lemko

Lemko will have Ukranian as its primary culture, as per what wikipedia
says Lemkos are. And polish is an accepted culture because they're in
poland.

* Removed Polish accepted culture from LMK

Upon further research, the peoples are not very similar at all and thus
the Lemkos shouldn't accept the Poles.

* Fixed the flags

Apparently Paint.NET doesn't save by file extension

* Made Lemko's color more intereseting

* LMK now has an upperhouse

tbh, I thought it would just be figured out automatically when released,
but apparently not.

* LMK parties now have names

They're just the Ukrainian party names, but with "Lemko" instead of
"Ukrainian"

* Updated the names of Lemko a little bit

They're not nearly as long now, and apparently Lemkovina is a word.

* Updated the flags for commie, fash, king. LMK

They aren't the best, but it was the best I could do!